Reorganized state query test code
