+++ Libplanet.Explorer.Tests/Queries/StateQueryTest.Legacy.cs
@@ -1,0 +1,344 @@
+using System.Collections.Generic;
+using System.Numerics;
+using System.Threading.Tasks;
+using GraphQL;
+using GraphQL.Execution;
+using Libplanet.Action.State;
+using Libplanet.Blockchain.Policies;
+using Libplanet.Explorer.Queries;
+using Libplanet.Types.Assets;
+using Xunit;
+using static Libplanet.Explorer.Tests.GraphQLTestUtils;
+
+namespace Libplanet.Explorer.Tests.Queries;
+
+public partial class StateQueryTest
+{
+    [Fact]
+    public async Task State()
+    {
+        (IBlockChainStates, IBlockPolicy) source = (
+            new MockChainStates(), new BlockPolicy()
+        );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            state(
+                 address: ""0x5003712B63baAB98094aD678EA2B24BcE445D076"",
+                 accountAddress: ""0x40837BFebC1b192600023a431400557EA5FDE51a""
+                 offsetBlockHash:
+                     ""01ba4719c80b6fe911b091a7c05124b64eeece964e09c058ef8f9805daca546b""
+            )
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        object state =
+            Assert.IsAssignableFrom<object>(resultDict["state"]);
+        Assert.Equal(new byte[] { 110, }, state);
+    }
+
+    [Fact]
+    public async Task Balance()
+    {
+        (IBlockChainStates, IBlockPolicy) source = (
+            new MockChainStates(), new BlockPolicy()
+        );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            balance(
+                 owner: ""0x5003712B63baAB98094aD678EA2B24BcE445D076"",
+                 currency: { ticker: ""ABC"", decimalPlaces: 2, totalSupplyTrackable: true },
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetBlockHash:
+                     ""01ba4719c80b6fe911b091a7c05124b64eeece964e09c058ef8f9805daca546b""
+            ) {
+                currency { ticker, hash }
+                sign
+                majorUnit
+                minorUnit
+                quantity
+                string
+            }
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        IDictionary<string, object> balanceDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultDict["balance"]);
+        IDictionary<string, object> currencyDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(balanceDict["currency"]);
+        Assert.Equal("ABC", currencyDict["ticker"]);
+        Assert.Equal("84ba810ca5ac342c122eb7ef455939a8a05d1d40", currencyDict["hash"]);
+        Assert.Equal(1, Assert.IsAssignableFrom<int>(balanceDict["sign"]));
+        Assert.Equal(123, Assert.IsAssignableFrom<BigInteger>(balanceDict["majorUnit"]));
+        Assert.Equal(0, Assert.IsAssignableFrom<BigInteger>(balanceDict["minorUnit"]));
+        Assert.Equal("123", balanceDict["quantity"]);
+        Assert.Equal("123 ABC", balanceDict["string"]);
+    }
+
+    [Fact]
+    public async Task TotalSupply()
+    {
+        var currency = Currency.Uncapped("ABC", 2, minters: null);
+#pragma warning disable CS0618  // Legacy, which is obsolete, is the only way to test this:
+        var legacyToken = Currency.Legacy("LEG", 0, null);
+#pragma warning restore CS0618
+        (IBlockChainStates, IBlockPolicy) source = (
+           new MockChainStates(), new BlockPolicy());
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            totalSupply(
+                 currency: { ticker: ""ABC"", decimalPlaces: 2, totalSupplyTrackable: true },
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetBlockHash:
+                     ""01ba4719c80b6fe911b091a7c05124b64eeece964e09c058ef8f9805daca546b""
+            ) {
+                currency { ticker, hash }
+                sign
+                majorUnit
+                minorUnit
+                quantity
+                string
+            }
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        IDictionary<string, object> totalSupplyDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultDict["totalSupply"]);
+        IDictionary<string, object> currencyDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(totalSupplyDict["currency"]);
+        Assert.Equal("ABC", currencyDict["ticker"]);
+        Assert.Equal("84ba810ca5ac342c122eb7ef455939a8a05d1d40", currencyDict["hash"]);
+        Assert.Equal(1, Assert.IsAssignableFrom<int>(totalSupplyDict["sign"]));
+        Assert.Equal(10000, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["majorUnit"]));
+        Assert.Equal(0, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["minorUnit"]));
+        Assert.Equal("10000", totalSupplyDict["quantity"]);
+        Assert.Equal("10000 ABC", totalSupplyDict["string"]);
+
+        result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            totalSupply(
+                 currency: { ticker: ""LEG"", decimalPlaces: 0, totalSupplyTrackable: false },
+                 offsetBlockHash:
+                     ""01ba4719c80b6fe911b091a7c05124b64eeece964e09c058ef8f9805daca546b""
+            ) {
+                quantity
+            }
+        }
+        ", source: source);
+        Assert.Single(result.Errors);
+        Assert.Contains("not trackable", result.Errors[0].Message);
+    }
+
+    [Fact]
+    public async Task Validators()
+    {
+        (IBlockChainStates, IBlockPolicy) source = (
+           new MockChainStates(),
+           new BlockPolicy()
+       );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            validators(
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetBlockHash:
+                     ""01ba4719c80b6fe911b091a7c05124b64eeece964e09c058ef8f9805daca546b""
+            ) {
+                publicKey
+                power
+            }
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        object[] validators = Assert.IsAssignableFrom<object[]>(resultDict["validators"]);
+        IDictionary<string, object> validatorDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(validators[0]);
+        Assert.Equal("032038e153d344773986c039ba5dbff12ae70cfdf6ea8beb7c5ea9b361a72a9233", validatorDict["publicKey"]);
+        Assert.Equal(new BigInteger(1), validatorDict["power"]);
+    }
+
+    [Fact]
+    public async Task ThrowExecutionErrorIfViolateMutualExclusive()
+    {
+        (IBlockChainStates, IBlockPolicy) source = (
+            new MockChainStates(), new BlockPolicy()
+        );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            states(
+                 addresses: [""0x5003712B63baAB98094aD678EA2B24BcE445D076"", ""0x0000000000000000000000000000000000000000""],
+                 offsetBlockHash:
+                     ""01ba4719c80b6fe911b091a7c05124b64eeece964e09c058ef8f9805daca546b"",
+                 offsetStateRootHash:
+                     ""c33b27773104f75ac9df5b0533854108bd498fab31e5236b6f1e1f6404d5ef64""
+            )
+        }
+        ", source: source);
+        Assert.IsType<ExecutionErrors>(result.Errors);
+    }
+
+    [Fact]
+    public async Task StateBySrh()
+    {
+        var currency = Currency.Uncapped("ABC", 2, minters: null);
+        (IBlockChainStates, IBlockPolicy) source = (
+            new MockChainStates(), new BlockPolicy()
+        );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            state(
+                 address: ""0x5003712B63baAB98094aD678EA2B24BcE445D076"",
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetStateRootHash:
+                     ""c33b27773104f75ac9df5b0533854108bd498fab31e5236b6f1e1f6404d5ef64""
+            )
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        object state =
+            Assert.IsAssignableFrom<object>(resultDict["state"]);
+        Assert.Equal(new byte[] { 110, }, state);
+    }
+
+    [Fact]
+    public async Task BalanceBySrh()
+    {
+        (IBlockChainStates, IBlockPolicy) source = (
+            new MockChainStates(), new BlockPolicy()
+        );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            balance(
+                 owner: ""0x5003712B63baAB98094aD678EA2B24BcE445D076"",
+                 currency: { ticker: ""ABC"", decimalPlaces: 2, totalSupplyTrackable: true },
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetStateRootHash:
+                     ""c33b27773104f75ac9df5b0533854108bd498fab31e5236b6f1e1f6404d5ef64""
+            ) {
+                currency { ticker, hash }
+                sign
+                majorUnit
+                minorUnit
+                quantity
+                string
+            }
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        IDictionary<string, object> balanceDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultDict["balance"]);
+        IDictionary<string, object> currencyDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(balanceDict["currency"]);
+        Assert.Equal("ABC", currencyDict["ticker"]);
+        Assert.Equal("84ba810ca5ac342c122eb7ef455939a8a05d1d40", currencyDict["hash"]);
+        Assert.Equal(1, Assert.IsAssignableFrom<int>(balanceDict["sign"]));
+        Assert.Equal(123, Assert.IsAssignableFrom<BigInteger>(balanceDict["majorUnit"]));
+        Assert.Equal(0, Assert.IsAssignableFrom<BigInteger>(balanceDict["minorUnit"]));
+        Assert.Equal("123", balanceDict["quantity"]);
+        Assert.Equal("123 ABC", balanceDict["string"]);
+    }
+
+    [Fact]
+    public async Task TotalSupplyBySrh()
+    {
+         var currency = Currency.Uncapped("ABC", 2, minters: null);
+#pragma warning disable CS0618  // Legacy, which is obsolete, is the only way to test this:
+         var legacyToken = Currency.Legacy("LEG", 0, null);
+#pragma warning restore CS0618
+        (IBlockChainStates, IBlockPolicy) source = (
+            new MockChainStates(), new BlockPolicy()
+        );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            totalSupply(
+                 currency: { ticker: ""ABC"", decimalPlaces: 2, totalSupplyTrackable: true },
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetStateRootHash:
+                     ""c33b27773104f75ac9df5b0533854108bd498fab31e5236b6f1e1f6404d5ef64""
+            ) {
+                currency { ticker, hash }
+                sign
+                majorUnit
+                minorUnit
+                quantity
+                string
+            }
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        IDictionary<string, object> totalSupplyDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultDict["totalSupply"]);
+        IDictionary<string, object> currencyDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(totalSupplyDict["currency"]);
+        Assert.Equal("ABC", currencyDict["ticker"]);
+        Assert.Equal("84ba810ca5ac342c122eb7ef455939a8a05d1d40", currencyDict["hash"]);
+        Assert.Equal(1, Assert.IsAssignableFrom<int>(totalSupplyDict["sign"]));
+        Assert.Equal(10000, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["majorUnit"]));
+        Assert.Equal(0, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["minorUnit"]));
+        Assert.Equal("10000", totalSupplyDict["quantity"]);
+        Assert.Equal("10000 ABC", totalSupplyDict["string"]);
+
+        result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            totalSupply(
+                 currency: { ticker: ""LEG"", decimalPlaces: 0, totalSupplyTrackable: false },
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetBlockHash:
+                     ""01ba4719c80b6fe911b091a7c05124b64eeece964e09c058ef8f9805daca546b""
+            ) {
+                quantity
+            }
+        }
+        ", source: source);
+        Assert.Single(result.Errors);
+        Assert.Contains("not trackable", result.Errors[0].Message);
+    }
+
+    [Fact]
+    public async Task ValidatorsBySrh()
+    {
+        (IBlockChainStates, IBlockPolicy) source = (
+            new MockChainStates(), new BlockPolicy()
+        );
+        ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
+        {
+            validators(
+                 accountAddress: ""0x1000000000000000000000000000000000000000"",
+                 offsetStateRootHash:
+                     ""c33b27773104f75ac9df5b0533854108bd498fab31e5236b6f1e1f6404d5ef64""
+            ) {
+                publicKey
+                power
+            }
+        }
+        ", source: source);
+        Assert.Null(result.Errors);
+        ExecutionNode resultData = Assert.IsAssignableFrom<ExecutionNode>(result.Data);
+        IDictionary<string, object> resultDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(resultData!.ToValue());
+        object[] validators = Assert.IsAssignableFrom<object[]>(resultDict["validators"]);
+        IDictionary<string, object> validatorDict =
+            Assert.IsAssignableFrom<IDictionary<string, object>>(validators[0]);
+        Assert.Equal("032038e153d344773986c039ba5dbff12ae70cfdf6ea8beb7c5ea9b361a72a9233", validatorDict["publicKey"]);
+        Assert.Equal(new BigInteger(1), validatorDict["power"]);
+    }
+}
